move transpose funciton into util file
