+++ src/daft-distributed/src/utils/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod channel;
 pub(crate) mod joinset;
 pub(crate) mod runtime;
 pub(crate) mod stream;
+pub(crate) mod transpose;
